datasets/data_filtering/image_information_filter/requirements.txt
numpy>=1.19.0
opencv-python-headless>=4.5.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.60.0
pillow>=8.2.0
webdataset>=0.2.0
pyarrow>=4.0.0
scipy>=1.7.0